telegram/requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gspread==6.2.1
httplib2==0.31.2
idna==3.11
oauth2client==4.1.3
oauthlib==3.3.1
pyaes==1.6.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
Telethon==1.43.1
urllib3==2.6.3
